+++ android/src/main/java/com/jhomlala/better_player/BetterPlayer.java
@@ -77,6 +77,7 @@ import java.util.Map;
 import com.google.android.exoplayer2.PlaybackParameters;
 
 final class BetterPlayer {
+    private static final String TAG = "BetterPlayer";
     private static final String FORMAT_SS = "ss";
     private static final String FORMAT_DASH = "dash";
     private static final String FORMAT_HLS = "hls";
@@ -664,50 +665,36 @@ final class BetterPlayer {
     }
 
 
-    void setAudio(String audioName, Context context) {
-        String TAG = "VIDEO_PLAYER_ANDROID";
-        Log.d(TAG, "SET AUDIO: " + audioName);
-        MappingTrackSelector.MappedTrackInfo mappedTrackInfo =
-                trackSelector.getCurrentMappedTrackInfo();
-
-        StringBuilder str = new StringBuilder();
-
-        for (int i = 0; i < mappedTrackInfo.getRendererCount(); i++) {
-            Log.d(TAG, "INDEX: " + i);
-
-            if (mappedTrackInfo.getRendererType(i) != C.TRACK_TYPE_AUDIO) {
-                Log.d(TAG, "NOT AUDIO. SKIPPED");
-                continue;
-            }
-
-            TrackGroupArray trackGroupArray = mappedTrackInfo.getTrackGroups(i);
-            for (int j = 0; j < trackGroupArray.length; j++) {
-                Log.d(TAG, "J:" + j);
-
-                TrackGroup group = trackGroupArray.get(j);
-
-                //TrackNameProvider provider = new DefaultTrackNameProvider(context.getResources());
-                for (int k = 0; k < group.length; k++) {
-                    Log.d(TAG, "K:" + k);
-                    Log.d(TAG, "FORMAT: " + group.getFormat(k).language);
-                    //Log.d(TAG,"TRACK NAME: >" + provider.getTrackName(group.getFormat(k))+"< vs >"+audioName+"<");
-                    if (group.getFormat(k).language.equals(audioName)) {
-                        Log.d(TAG, "USING TRACK NAME!!!" + audioName);
-
-                        DefaultTrackSelector.ParametersBuilder builder = trackSelector.getParameters().buildUpon();
-                        builder.clearSelectionOverrides(i).setRendererDisabled(i, false);
-                        int[] tracks = {k};
-                        DefaultTrackSelector.SelectionOverride override = new DefaultTrackSelector.SelectionOverride(j, tracks);
-                        builder.setSelectionOverride(i, mappedTrackInfo.getTrackGroups(i), override);
-                        trackSelector.setParameters(builder);
-                        return;
-
-
+    void setAudioTrack(String languageCode) {
+        try {
+            MappingTrackSelector.MappedTrackInfo mappedTrackInfo =
+                    trackSelector.getCurrentMappedTrackInfo();
+            if (mappedTrackInfo != null) {
+                for (int i = 0; i < mappedTrackInfo.getRendererCount(); i++) {
+                    if (mappedTrackInfo.getRendererType(i) != C.TRACK_TYPE_AUDIO) {
+                        continue;
                     }
+                    TrackGroupArray trackGroupArray = mappedTrackInfo.getTrackGroups(i);
+                    for (int j = 0; j < trackGroupArray.length; j++) {
+                        TrackGroup group = trackGroupArray.get(j);
+                        for (int k = 0; k < group.length; k++) {
+                            String groupLanguage = group.getFormat(k).language;
+                            if (groupLanguage != null && groupLanguage.equals(languageCode)) {
+                                DefaultTrackSelector.ParametersBuilder builder = trackSelector.getParameters().buildUpon();
+                                builder.clearSelectionOverrides(i).setRendererDisabled(i, false);
+                                int[] tracks = {k};
+                                DefaultTrackSelector.SelectionOverride override = new DefaultTrackSelector.SelectionOverride(j, tracks);
+                                builder.setSelectionOverride(i, mappedTrackInfo.getTrackGroups(i), override);
+                                trackSelector.setParameters(builder);
+                                return;
+                            }
 
+                        }
+                    }
                 }
             }
-
+        } catch (Exception exception) {
+            Log.e(TAG, "setAudioTrack failed" + exception.toString());
         }
     }