+++ lib/src/controls/better_player_cupertino_controls.dart
@@ -564,7 +564,7 @@ class _BetterPlayerCupertinoControlsState
     _updateState();
 
     if ((_controller.value != null && _controller.value.isPlaying) ||
-        _betterPlayerController.autoPlay) {
+        _betterPlayerController.betterPlayerConfiguration.autoPlay) {
       _startHideTimer();
     }
 
@@ -686,8 +686,10 @@ class _BetterPlayerCupertinoControlsState
   }
 
   Widget _buildErrorWidget() {
-    if (_betterPlayerController.errorBuilder != null) {
-      return _betterPlayerController.errorBuilder(context,
+    final errorBuilder =
+        _betterPlayerController.betterPlayerConfiguration.errorBuilder;
+    if (errorBuilder != null) {
+      return errorBuilder(context,
           _betterPlayerController.videoPlayerController.value.errorDescription);
     } else {
       final textStyle = TextStyle(color: _controlsConfiguration.textColor);